every goal can open by itself
